requirements.txt
fastapi
uvicorn[standard]
celery
redis
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
pydantic-settings
boto3
PyMuPDF